+++ src/util/misc/projectStroke/StrokeLineCapProjections.ts
@@ -1,6 +1,6 @@
 import type { XY } from '../../../Point';
 import { Point } from '../../../Point';
-import { createVector, getOrthonormalVector, getUnitVector } from '../vectors';
+import { getOrthonormalVector, getUnitVector } from '../vectors';
 import { StrokeLineJoinProjections } from './StrokeLineJoinProjections';
 import { StrokeProjectionsBase } from './StrokeProjectionsBase';
 import type { TProjection, TProjectStrokeOnPointsOptions } from './types';
@@ -62,12 +62,32 @@ export class StrokeLineCapProjections extends StrokeProjectionsBase {
    * @see https://github.com/fabricjs/fabric.js/pull/8344#1-2-round
    */
   projectRound() {
-    return new StrokeLineJoinProjections(
-      this.A,
-      this.T,
-      this.T,
-      this.options
-    ).projectRound();
+    const projections: Point[] = [];
+
+    if (!this.isSkewed() && this.A.eq(this.T)) {
+      /* 1 point case without `skew`
+        When `strokeUniform` is true, scaling has no effect.
+        So we divide by scale, to remove its effect.
+      */
+      const projection = new Point(1, 1)
+        .scalarMultiply(this.strokeProjectionMagnitude)
+        .multiply(this.strokeUniformScalar);
+      projections.push(
+        this.applySkew(this.A.add(projection)),
+        this.applySkew(this.A.subtract(projection))
+      );
+    } else {
+      projections.push(
+        ...new StrokeLineJoinProjections(
+          this.A,
+          this.T,
+          this.T,
+          this.options
+        ).projectRound()
+      );
+    }
+
+    return projections;
   }
 
   /**
@@ -77,20 +97,35 @@ export class StrokeLineCapProjections extends StrokeProjectionsBase {
    * @see https://github.com/fabricjs/fabric.js/pull/8344#1-3-square
    */
   projectSquare() {
-    const orthogonalProjection = this.calcOrthogonalProjection(
-      this.A,
-      this.T,
-      this.strokeProjectionMagnitude
-    );
-    const strokePointingOut = this.scaleUnitVector(
-      getUnitVector(createVector(this.A, this.T)),
-      -this.strokeProjectionMagnitude
-    );
-    const projectedA = this.A.add(strokePointingOut);
-    return [
-      projectedA.add(orthogonalProjection),
-      projectedA.subtract(orthogonalProjection),
-    ].map((p) => this.applySkew(p));
+    const projections: Point[] = [];
+
+    if (this.A.eq(this.T)) {
+      /* 1 point case without `skew`
+        When `strokeUniform` is true, scaling has no effect.
+        So we divide by scale, to remove its effect.
+      */
+      const projection = new Point(1, 1)
+        .scalarMultiply(this.strokeProjectionMagnitude)
+        .multiply(this.strokeUniformScalar);
+      projections.push(this.A.add(projection), this.A.subtract(projection));
+    } else {
+      const orthogonalProjection = this.calcOrthogonalProjection(
+        this.A,
+        this.T,
+        this.strokeProjectionMagnitude
+      );
+      const strokePointingOut = this.scaleUnitVector(
+        getUnitVector(this.createSideVector(this.A, this.T)),
+        -this.strokeProjectionMagnitude
+      );
+      const projectedA = this.A.add(strokePointingOut);
+      projections.push(
+        projectedA.add(orthogonalProjection),
+        projectedA.subtract(orthogonalProjection)
+      );
+    }
+
+    return projections.map((p) => this.applySkew(p));
   }
 
   protected projectPoints() {